fix: captcha token in password service
