+++ packages/apps/human-app/frontend/src/modules/worker/reset-password/reset-password.page.tsx
@@ -15,11 +15,9 @@ import { getErrorMessageForError } from '@/shared/errors';
 import { routerPaths } from '@/router/router-paths';
 import { HCaptchaForm } from '@/shared/components/hcaptcha';
 import { useResetMutationErrors } from '@/shared/hooks/use-reset-mutation-errors';
-import {
-  type ResetPasswordDto,
-  resetPasswordDtoSchema,
-  useResetPasswordMutation,
-} from './hooks';
+import { useResetPasswordMutation } from './hooks';
+import { resetPasswordDtoSchema } from './schemas';
+import { type ResetPasswordDto } from './types';
 
 export function ResetPasswordWorkerPage() {
   const location = useLocation();
