+++ packages/apps/human-app/frontend/src/modules/worker/reset-password/types.ts
@@ -1,0 +1,4 @@
+import { type z } from 'zod';
+import { type resetPasswordDtoSchema } from './schemas';
+
+export type ResetPasswordDto = z.infer<typeof resetPasswordDtoSchema>;
